schema: network protocol enum type
