+++ validate/fixtures/firewall.in.yaml
@@ -6,11 +6,11 @@ network:
       - service: "dhcp"
         enabled: false
       - service: "ssh"
-        protocol: "tcp"
+        protocol: tcp
     ports:
       - port: 53
         protocol: "any"
         enabled: false
       - from: 60000
         to: 60010
-        protocol: "udp"
+        protocol: udp
